requirements.txt
streamlit
pillow
numpy
requests
qrcode
pyzbar
supabase
